Begining the refactor of Mr Bones.
The goal here is twofold. First, I want others to be able to develop their own
functinoality for Mr Bones. To that end I'm incorporating the little-plugger
plugin system for functionality and the loquactions configuration system for
configuration and help documentation.

Here's to not screwing stuff up along the way :)