More informative error message when gems cannot be found.
If the user attempts to configure some rcov options and the rcov gem is not
installed, then we'll get an error message "undefined method for nil". This
happens because bones will not load the configuration options for gems that
are not installed. So, if the rcov gem is missing, then there will be no
"rcov" configuration option. Referencing something like "rcov.opts" will not
work because "rcov" returns nil, and "rcov.opts" raises an error with no
method "opts" defined for nil.