Use git rules for lines in that start with `/` in .gitignore
In `.gitignore` a rule like

    /dir/

makes git ignore only the directory named `dir` that is in the top
level, not all the directories named `dir`. While it would be hard for
Bones to have exactly the same behaviour, this patch simulates how git
would behave in the case of a `/dir/` line.