Don't import rdoc options from gemspec if rdoc plugin isn't loaded
If we choose to disregard the rdoc plugin and we tell Mr Bones to import an
existing gemspec, then Loquacious will display a warning like this:

  ---------------------------------------------------------------------------
  The Loquacious configuration system has detected that one or moe
  settings have an undefined value. An attempt is being made to reference
  sub-properties of these undefined settings. Messages will follow containing
  information about the undefined properties.
  ---------------------------------------------------------------------------
  Access to undefined value "rdoc": rdoc.opts=

This is because the rdoc plugin adds rdoc.opts to the global config object. So,
we merely check that the rdoc plugin has been loaded before attempting to set
this option.